requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
prometheus-client>=0.19.0
pydantic>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-exporter-gcp-monitoring>=1.11.0a0
opentelemetry-exporter-gcp-trace>=1.11.0a0
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
